requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
idna==2.8
python-gitlab==1.9.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
